requirements.txt
Flask==3.0.0
Pillow==10.1.0
rembg==2.0.52
onnxruntime==1.16.3
python-multipart
gunicorn
